4.21

+++ controllers/blogs.js
@@ -55,12 +55,29 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     try {
-        await Blog.findByIdAndRemove(request.params.id)
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const userid = decodedToken.id
+        const blog = await Blog.findById(request.params.id)
+
+        if (blog.user.toString() === userid.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            return response.status(401).json({ error: 'invalid user' })
+        }
     
-        response.status(204).end()
-      } catch (exception) {
-        console.log(exception)
-        response.status(400).send({ error: 'malformatted id' })
+    } catch(exception) {
+        if (exception.name === 'JsonWebTokenError' ) {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(500).json({ error: 'something went wrong...' })
+        }
     }
 })
